--- ydb/core/persqueue/config/ya.make
@@ -1,12 +1,0 @@
-LIBRARY()
-
-SRCS(
-    config.h
-    config.cpp
-)
-
-PEERDIR(
-    ydb/core/protos
-)
-
-END()
