+++ ydb/core/persqueue/public/ya.make
@@ -1,6 +1,7 @@
 LIBRARY()
 
 SRCS(
+    config.cpp
     pq_database.cpp
     pq_rl_helpers.cpp
     utils.cpp
