+++ ydb/core/tx/schemeshard/ya.make
@@ -345,7 +345,6 @@ PEERDIR(
     ydb/core/keyvalue
     ydb/core/keyvalue/protos
     ydb/core/metering
-    ydb/core/persqueue/config
     ydb/core/persqueue/events
     ydb/core/persqueue/public
     ydb/core/persqueue/public/partition_index_generator
